requirements.txt
fastapi==0.112.1
praw==7.7.1
python-dotenv==1.0.1
uvicorn==0.30.6
